requirements.txt
asgiref==3.3.1
Django==3.1.7
django-test-plus==1.4.0
djangorestframework==3.12.4
psycopg2-binary==2.8.6
sqlparse==0.4.1
